🌟 ✨ MVP of News Centre custom module, yay! 😃
First attempt at a custom Drupal 8 module to render the "News Centre" page on the Intranet.

This version just grabs the Announcements RSS feed from the Commerce website, parses the XML into an array and uses Twig to spit out
the titles and URLs of the latest announcements into a simple list.